D3ASIM-3348: Refreshing cn_device_mapping after 15mins for both mqtt and ws.
