+++ live_data_subscriber/constants.py
@@ -1,0 +1,4 @@
+MEASUREMENT_PERIOD_MINUTES = 15
+RELOAD_CN_DEVICE_LIST_TIMEOUT_SECONDS = 60 * MEASUREMENT_PERIOD_MINUTES
+MINUTES_IN_HOUR = 60
+MQTT_PORT = 1883
